Move protocol folder to the root directory
